pyproject.toml
[tool.poetry]
name = "nonebot-plugin-fun-content"
version = "0.2.3"
description = "从多个 API 获取趣味内容，如一言、土味情话、舔狗日记等。"
authors = ["chsiyu <rk.chensiyu@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.25.0"
nonebot2 = "^2.2.0"
nonebot-adapter-onebot = "^2.3.0"
pydantic = "^1.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"